LG-11285: Reset the callback property descriptor on cleanup of useEffect. (#10762)
* Extract defineObservableProperty into its own file
* Add stopObservingProperty
* Use stopObservingProperty in acuant-capture-canvas

changelog: Bug Fixes, Acuant SDK, Clear out a callback when done with it.
---------
